awful.rules: Fix `floating` and `size_hints_honor` with functions
Most properties in rules accept functions instead of plain values; in
this case the actual value of the property is determined by calling the
function.  However, the expressions used for the `floating` and
`size_hints_honor` properties were not giving the correct result if the
function returned `false` or `nil` (in this case the code attempted to
use the function itself as the property value, which resulted in Lua
errors).  Replace the expressions with conditional operators.

Signed-off-by: Sergey Vlasov <sigprof@gmail.com>