test-awful-placement: Make the test repeatable
In order to test the behavior of `awful.placement.no_overlap` with
unselected tags, the test sequence must be able to run multiple times.
Fix the test code to make this possible (currently it just performs the
same sequence 3 times, the code to actually test the behavior with
different tags will be added later).

Signed-off-by: Sergey Vlasov <sigprof@gmail.com>