awful.placement: Fix no_overlap with unselected tags (#2809)
The awful.placement.no_overlap function always looked at the currently
visible clients when placing a new client.  This produced a confusing
result when using awful.rules or the sn_rules argument of awful.spawn to
place the client on an unselected tag (the client was placed as if it
would be placed on a currently selected tag; if multiple clients were
placed on the same unselected tag, in many cases they were placed at the
same position, overlapping each other).

Make awful.placement.no_overlap check tags of the placed client and
handle the case of placement on an unselected tag in a more useful way:

  - If at least one of the client tags is selected, keep the previous
    behavior: avoid overlap with all other floating clients which are
    currently visible, and use the currently active layout to determine
    the floating status.

  - If all client tags are unselected, avoid overlap with all other
    clients which share at least one tag with the placed client, and use
    the layout of the first tag of the placed client to determine the
    floating status.

Signed-off-by: Sergey Vlasov <sigprof@gmail.com>